Fix missing space before appended text
Caused by abstracts code stripping leading spaces (and trailing ones)
from lines as they're loaded. This means when you first rehash and
immediately test it looks like it works, but after the abstract has
expired and is reloaded it'll be broken.

Fix the "%space" replacement code to operate on all instances on a line.

Also fixed some indentation/spacing.